+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## dev
+
+* Enhancements
+  * [Installation] Installation script (`install.sh`) now installs Docker automatically;
+
 ## v0.16.2 - (2015-11-17)
 
 * Enhancements